Fix IR code generation bugs for self-hosting
Fix a chain of codegen issues that prevented the self-hosted
compiler from running:

- Promote/narrow integer call args to match declared parameter
  types (fixes isalpha(i8) instead of isalpha(i32) causing the
  preprocessor hang).
- Add variadic function support: emit '...' in declare and apply
  default argument promotion (char->int, float->double).
- Add switch-statement dispatch and multi-statement case bodies.
- Fix 2D-array GEP to use the full array type as base.
- Fix address-of (&) for lvalues to return the address via
  gen_store_ptr, not the value.
- Fix sizeof(array) to use the array type (no pointer decay).
- Add scoped symbol table so shadowed locals are restored on
  block exit.
- Widen int_val to long long and parse ULL literals correctly
  (fixes FNV hash constants being truncated).
- Unify anonymous structs by members pointer in the IR dumper.
- Increase MAX_PRE for switch merge blocks.
- Add local array initializer emission and guard integer
  widening in ternary/binary coercion.

+++ tokenizer/token.h
@@ -104,7 +104,7 @@ struct Token {
     Token*    next;
 
     union {
-        long int_val;           // TOK_INT_LIT
+        long long int_val;      // TOK_INT_LIT / TOK_LONG_LIT (64-bit)
         char char_val;          // TOK_CHAR_LIT
         double float_val;       // TOK_FLOAT_LIT
         String str_val;              // TOK_STRING_LIT
